chg: dev: massive code refactor

+++ stubs/nginx_servervars.go
@@ -1,8 +1,10 @@
 package stubs
 
+// ServerVars represents the values to replace when parsing an NGINX Server
+// template
 type ServerVars struct {
-	Port       uint
-	SslPort    uint
+	Port       uint16
+	SslPort    uint16
 	ServerName string
 	ServerRoot string
 }
